Change trellis-auth-oauth to trellis-oauth
