+++ auth/oauth/src/test/java/org/trellisldp/oauth/JwksAuthenticatorTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.trellisldp.auth.oauth;
+package org.trellisldp.oauth;
 
 import static java.time.Instant.now;
 import static java.util.Base64.getUrlDecoder;
