+++ components/dropwizard/src/main/java/org/trellisldp/dropwizard/TrellisUtils.java
@@ -31,15 +31,15 @@ import javax.ws.rs.container.ContainerRequestFilter;
 import org.trellisldp.api.ResourceService;
 import org.trellisldp.api.RuntimeTrellisException;
 import org.trellisldp.auth.basic.BasicAuthFilter;
-import org.trellisldp.auth.oauth.Authenticator;
-import org.trellisldp.auth.oauth.NullAuthenticator;
-import org.trellisldp.auth.oauth.OAuthFilter;
-import org.trellisldp.auth.oauth.OAuthUtils;
 import org.trellisldp.cache.TrellisCache;
 import org.trellisldp.dropwizard.config.AuthConfiguration;
 import org.trellisldp.dropwizard.config.CORSConfiguration;
 import org.trellisldp.dropwizard.config.JwtAuthConfiguration;
 import org.trellisldp.dropwizard.config.TrellisConfiguration;
+import org.trellisldp.oauth.Authenticator;
+import org.trellisldp.oauth.NullAuthenticator;
+import org.trellisldp.oauth.OAuthFilter;
+import org.trellisldp.oauth.OAuthUtils;
 import org.trellisldp.webac.AuthorizedModes;
 import org.trellisldp.webac.WebAcService;
