+++ platform/openliberty/build.gradle
@@ -18,8 +18,6 @@ dependencies {
     implementation project(':trellis-api')
     implementation project(':trellis-app')
     implementation project(':trellis-audit')
-    implementation project(':trellis-auth-basic')
-    implementation project(':trellis-auth-oauth')
     implementation project(':trellis-cdi')
     implementation project(':trellis-constraint-rules')
     implementation project(':trellis-event-jsonb')
@@ -27,6 +25,7 @@ dependencies {
     implementation project(':trellis-http')
     implementation project(':trellis-io-jena')
     implementation project(':trellis-namespaces')
+    implementation project(':trellis-oauth')
     implementation project(':trellis-rdfa')
     implementation project(':trellis-triplestore')
     implementation project(':trellis-vocabulary')
